refactor: move paginated data to data/ and other changes
