+++ src/apps/members/apps/segments/app.ts
@@ -3,7 +3,7 @@ import { getRepository } from "typeorm";
 
 import { hasNewModel } from "@core/middleware";
 import { wrapAsync } from "@core/utils";
-import { buildQuery } from "@core/utils/rules";
+import { buildQuery } from "@core/utils/member";
 
 import SegmentService from "@core/services/SegmentService";
